Move calculation function to a separate unit
